Desacoplamiento de resoponsabilidades
